requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile 
#
arrow==0.13.1             # via rq-dashboard
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via pytest
certifi==2019.3.9         # via requests
chardet==3.0.4            # via requests
click==7.0                # via clickclick, flask, rq
clickclick==1.2.2         # via connexion
connexion[swagger-ui]==2.2.0
filelock==3.0.10          # via tox
flask==1.0.2              # via connexion, rq-dashboard
gevent==1.4.0
greenlet==0.4.15          # via gevent
gunicorn==19.9.0
idna==2.8                 # via requests
inflection==0.3.1         # via connexion
itsdangerous==1.1.0       # via flask
jinja2==2.10.1            # via flask, swagger-ui-bundle
jsonschema==2.6.0         # via connexion, openapi-spec-validator
markupsafe==1.1.1         # via jinja2
more-itertools==7.0.0     # via pytest
openapi-spec-validator==0.2.6  # via connexion
pathlib==1.0.1            # via openapi-spec-validator
pluggy==0.9.0             # via pytest, tox
py==1.8.0                 # via pytest, tox
pytest==4.4.1
python-dateutil==2.8.0    # via arrow
pyyaml==5.1               # via clickclick, connexion, openapi-spec-validator
redis==3.2.1
requests==2.21.0          # via connexion
rq-dashboard==0.4.0
rq==0.13.0
six==1.12.0               # via connexion, openapi-spec-validator, pytest, python-dateutil, tox
swagger-ui-bundle==0.0.3  # via connexion
toml==0.10.0              # via tox
tox==3.9.0
urllib3==1.24.2           # via requests
virtualenv==16.5.0        # via tox
werkzeug==0.15.2          # via flask
